Revert "[serve] Support `working_dir` in `serve run` (#22760)"
This reverts commit ab2741d64ba029ecb53b66fa7c5b1509d551e12f.
