+++ dashboard/modules/job/tests/test_sdk.py
@@ -2,7 +2,7 @@ import pytest
 from typing import Dict, Optional, Tuple
 from unittest.mock import Mock, patch
 
-from ray.dashboard.modules.dashboard_sdk import parse_cluster_info
+from ray.dashboard.modules.job.sdk import parse_cluster_info
 
 
 @pytest.mark.parametrize(
@@ -25,12 +25,12 @@ def test_parse_cluster_info(
     headers: Optional[Dict[str, str]],
 ):
     """
-    Test ray.dashboard.modules.dashboard_sdk.parse_cluster_info for different
+    Test ray.dashboard.modules.job.sdk.parse_cluster_info for different
     format of addresses.
     """
-    mock_get_submission_client_cluster = Mock(return_value="Ray ClusterInfo")
+    mock_get_job_submission_client_cluster = Mock(return_value="Ray ClusterInfo")
     mock_module = Mock()
-    mock_module.get_submission_client_cluster_info = Mock(
+    mock_module.get_job_submission_client_cluster_info = Mock(
         return_value="Other module ClusterInfo"
     )
     mock_import_module = Mock(return_value=mock_module)
@@ -38,8 +38,8 @@ def test_parse_cluster_info(
     address, module_string, inner_address = address_param
 
     with patch.multiple(
-        "ray.dashboard.modules.dashboard_sdk",
-        get_submission_client_cluster_info=mock_get_submission_client_cluster,
+        "ray.dashboard.modules.job.sdk",
+        get_job_submission_client_cluster_info=mock_get_job_submission_client_cluster,
     ), patch.multiple("importlib", import_module=mock_import_module):
         if module_string == "ray":
             assert (
@@ -52,7 +52,7 @@ def test_parse_cluster_info(
                 )
                 == "Ray ClusterInfo"
             )
-            mock_get_submission_client_cluster.assert_called_once_with(
+            mock_get_job_submission_client_cluster.assert_called_once_with(
                 inner_address,
                 create_cluster_if_needed=create_cluster_if_needed,
                 cookies=cookies,
@@ -71,7 +71,7 @@ def test_parse_cluster_info(
                 == "Other module ClusterInfo"
             )
             mock_import_module.assert_called_once_with(module_string)
-            mock_module.get_submission_client_cluster_info.assert_called_once_with(
+            mock_module.get_job_submission_client_cluster_info.assert_called_once_with(
                 inner_address,
                 create_cluster_if_needed=create_cluster_if_needed,
                 cookies=cookies,
