+++ python/ray/serve/scripts.py
@@ -4,6 +4,7 @@ import yaml
 import os
 import sys
 import pathlib
+import requests
 import click
 import time
 from typing import Tuple, List, Dict
@@ -24,9 +25,18 @@ from ray.dashboard.modules.serve.schema import (
     schema_to_serve_application,
     serve_application_status_to_schema,
 )
-from ray.dashboard.modules.dashboard_sdk import parse_runtime_env_args
-from ray.dashboard.modules.serve.sdk import ServeSubmissionClient
 from ray.autoscaler._private.cli_logger import cli_logger
+
+
+def log_failed_request(response: requests.models.Response, address: str):
+    error_message = (
+        f"\nRequest to address {address} failed. Got response status code "
+        f"{response.status_code} with the following message:"
+        f"\n\n{response.text}"
+    )
+    cli_logger.newline()
+    cli_logger.error(error_message)
+    cli_logger.newline()
 
 
 def process_args_and_kwargs(
@@ -89,33 +99,14 @@ def process_args_and_kwargs(
     return args, args_and_kwargs
 
 
-def configure_runtime_env(deployment: Deployment, updates: Dict):
-    """
-    Overwrites deployment's runtime_env with fields in updates. Any fields in
-    deployment's runtime_env that aren't in updates stay the same.
-    """
-
-    if deployment.ray_actor_options is None:
-        deployment._ray_actor_options = {"runtime_env": updates}
-    elif "runtime_env" in deployment.ray_actor_options:
-        deployment.ray_actor_options["runtime_env"].update(updates)
-    else:
-        deployment.ray_actor_options["runtime_env"] = updates
-
-
 @click.group(help="[EXPERIMENTAL] CLI for managing Serve instances on a Ray cluster.")
-def cli():
-    pass
-
-
-@cli.command(help="Start a detached Serve instance on the Ray cluster.")
 @click.option(
     "--address",
     "-a",
     default=os.environ.get("RAY_ADDRESS", "auto"),
     required=False,
     type=str,
-    help='Address of the running Ray cluster to connect to. Defaults to "auto".',
+    help="Address of the running Ray cluster to connect to. " 'Defaults to "auto".',
 )
 @click.option(
     "--namespace",
@@ -132,6 +123,15 @@ def cli():
     type=str,
     help=("Runtime environment dictionary to pass into ray.init. Defaults to empty."),
 )
+def cli(address, namespace, runtime_env_json):
+    ray.init(
+        address=address,
+        namespace=namespace,
+        runtime_env=json.loads(runtime_env_json),
+    )
+
+
+@cli.command(help="Start a detached Serve instance on the Ray cluster.")
 @click.option(
     "--http-host",
     default=DEFAULT_HTTP_HOST,
@@ -160,20 +160,7 @@ def cli():
     type=str,
     hidden=True,
 )
-def start(
-    address,
-    namespace,
-    runtime_env_json,
-    http_host,
-    http_port,
-    http_location,
-    checkpoint_path,
-):
-    ray.init(
-        address=address,
-        namespace=namespace,
-        runtime_env=json.loads(runtime_env_json),
-    )
+def start(http_host, http_port, http_location, checkpoint_path):
     serve.start(
         detached=True,
         http_options=dict(
@@ -186,27 +173,7 @@ def start(
 
 
 @cli.command(help="Shutdown the running Serve instance on the Ray cluster.")
-@click.option(
-    "--address",
-    "-a",
-    default=os.environ.get("RAY_ADDRESS", "auto"),
-    required=False,
-    type=str,
-    help='Address of the running Ray cluster to connect to. Defaults to "auto".',
-)
-@click.option(
-    "--namespace",
-    "-n",
-    default="serve",
-    required=False,
-    type=str,
-    help='Ray namespace to connect to. Defaults to "serve".',
-)
-def shutdown(address: str, namespace: str):
-    ray.init(
-        address=address,
-        namespace=namespace,
-    )
+def shutdown():
     serve.api._connect()
     serve.shutdown()
 
@@ -222,47 +189,13 @@ class may or may not be decorated with ``@serve.deployment``.
 )
 @click.argument("deployment")
 @click.option(
-    "--address",
-    "-a",
-    default=os.environ.get("RAY_ADDRESS", "auto"),
-    required=False,
-    type=str,
-    help='Address of the running Ray cluster to connect to. Defaults to "auto".',
-)
-@click.option(
-    "--namespace",
-    "-n",
-    default="serve",
-    required=False,
-    type=str,
-    help='Ray namespace to connect to. Defaults to "serve".',
-)
-@click.option(
-    "--runtime-env-json",
-    default=r"{}",
-    required=False,
-    type=str,
-    help=("Runtime environment dictionary to pass into ray.init. Defaults to empty."),
-)
-@click.option(
     "--options-json",
     default=r"{}",
     required=False,
     type=str,
     help="JSON string for the deployments options",
 )
-def create_deployment(
-    address: str,
-    namespace: str,
-    runtime_env_json: str,
-    deployment: str,
-    options_json: str,
-):
-    ray.init(
-        address=address,
-        namespace=namespace,
-        runtime_env=json.loads(runtime_env_json),
-    )
+def create_deployment(deployment: str, options_json: str):
     deployment_cls = import_attr(deployment)
     if not isinstance(deployment_cls, Deployment):
         deployment_cls = serve.deployment(deployment_cls)
@@ -271,17 +204,11 @@ def create_deployment(
 
 
 @cli.command(
-    short_help="[Experimental] Deploy deployments from a YAML config file.",
-    help=(
-        "Deploys deployment(s) from CONFIG_OR_IMPORT_PATH, which must be either a "
-        "Serve YAML configuration file path or an import path to "
-        "a class or function to deploy.\n\n"
-        "Import paths must be of the form "
-        '"module.submodule_1...submodule_n.MyClassOrFunction".\n\n'
-        "Sends a nonblocking request. A successful response only indicates that the "
-        "request was received successfully. It does not mean the deployments are "
-        "live. Use `serve info` and `serve status` to check on them. "
-    ),
+    help="""
+    [Experimental] Deploy a YAML configuration file via REST API to
+    your Serve cluster.
+    """,
+    hidden=True,
 )
 @click.argument("config_file_name")
 @click.option(
@@ -293,111 +220,51 @@ def create_deployment(
     help='Address of the Ray dashboard to query. For example, "http://localhost:8265".',
 )
 def deploy(config_file_name: str, address: str):
+    full_address_path = f"{address}/api/serve/deployments/"
 
     with open(config_file_name, "r") as config_file:
         config = yaml.safe_load(config_file)
 
-    # Schematize config to validate format
+    # Generate a schema using the config to ensure its format is valid
     ServeApplicationSchema.parse_obj(config)
 
-    ServeSubmissionClient(address).deploy_application(config)
+    response = requests.put(full_address_path, json=config)
 
-    cli_logger.newline()
-    cli_logger.success(
-        "\nSent deploy request successfully!\n "
-        "* Use `serve status` to check your deployments' statuses.\n "
-        "* Use `serve info` to see your running Serve "
-        "application's configuration.\n"
-    )
-    cli_logger.newline()
+    if response.status_code == 200:
+        cli_logger.newline()
+        cli_logger.success(
+            "\nSent deploy request successfully!\n "
+            "* Use `serve status` to check your deployments' statuses.\n "
+            "* Use `serve info` to see your running Serve "
+            "application's configuration.\n"
+        )
+        cli_logger.newline()
+    else:
+        log_failed_request(response, address)
 
 
 @cli.command(
-    short_help="[Experimental] Run deployments via Serve's Python API.",
-    help=(
-        "Deploys deployment(s) from CONFIG_OR_IMPORT_PATH, which must be either a "
-        "Serve YAML configuration file path or an import path to "
-        "a class or function to deploy.\n\n"
-        "The full command must be of the form:\n"
-        '"serve run [import path] [optional parameters] -- [arg-1] ... [arg-n] '
-        '[kwarg-1]=[kwval-1] ... [kwarg-n]=[kwval-n]"\n\n'
-        "Deployments via import path may also take in init_args and "
-        "init_kwargs from any ARGS_AND_KWARGS passed in. Import paths must be "
-        "of the form:\n"
-        '"module.submodule_1...submodule_n.MyClassOrFunction".\n\n'
-        "Blocks after deploying, and logs status periodically. After being killed, "
-        "this command tears down all deployments it deployed. If there are no "
-        "deployments left, it also tears down the Serve application."
-    ),
+    help="[Experimental] Run deployments via Serve's Python API.",
+    hidden=True,
 )
 @click.argument("config_or_import_path")
 @click.argument("args_and_kwargs", required=False, nargs=-1)
-@click.option(
-    "--runtime-env",
-    type=str,
-    default=None,
-    required=False,
-    help="Path to a local YAML file containing a runtime_env definition. "
-    "Overrides all runtime_envs specified in a config file.",
-)
-@click.option(
-    "--runtime-env-json",
-    type=str,
-    default=None,
-    required=False,
-    help="JSON-serialized runtime_env dictionary. Overrides all runtime_envs "
-    "specified in a config file.",
-)
-@click.option(
-    "--working-dir",
-    type=str,
-    default=None,
-    required=False,
-    help=(
-        "Directory containing files that your job will run in. Can be a "
-        "local directory or a remote URI to a .zip file (S3, GS, HTTP). "
-        "This overrides the working_dir in --runtime-env if both are "
-        "specified. Overrides all working_dirs specified in a config file."
-    ),
-)
-@click.option(
-    "--cluster-address",
-    "-c",
-    default="auto",
-    required=False,
-    type=str,
-    help=('Address of the Ray cluster to query. Defaults to "auto".'),
-)
-@click.option(
-    "--dashboard-address",
-    "-d",
-    default="http://localhost:8265",
-    required=False,
-    type=str,
-    help=(
-        'Address of the Ray dashboard to query. Defaults to "http://localhost:8265".'
-    ),
-)
 def run(
     config_or_import_path: str,
     args_and_kwargs: Tuple[str],
-    runtime_env: str,
-    runtime_env_json: str,
-    working_dir: str,
-    cluster_address: str,
-    dashboard_address: str,
 ):
+    """
+    Deploys deployment(s) from CONFIG_OR_IMPORT_PATH, which must be either a
+    Serve YAML configuration file path or an import path to
+    a class or function to deploy. Import paths must be of the form
+    "module.submodule_1...submodule_n.MyClassOrFunction".
+    """
 
     try:
         # Check if path provided is for config or import
         deployments = []
         is_config = pathlib.Path(config_or_import_path).is_file()
         args, kwargs = process_args_and_kwargs(args_and_kwargs)
-        runtime_env_updates = parse_runtime_env_args(
-            runtime_env=runtime_env,
-            runtime_env_json=runtime_env_json,
-            working_dir=working_dir,
-        )
 
         if is_config:
             config_path = config_or_import_path
@@ -418,14 +285,7 @@ def run(
             schematized_config = ServeApplicationSchema.parse_obj(config)
             deployments = schema_to_serve_application(schematized_config)
 
-            ray.init(address=cluster_address, namespace="serve")
             serve.start(detached=True)
-            ServeSubmissionClient(dashboard_address)._upload_working_dir_if_needed(
-                runtime_env_updates
-            )
-
-            for deployment in deployments:
-                configure_runtime_env(deployment, runtime_env_updates)
             deploy_group(deployments)
 
             cli_logger.newline()
@@ -450,13 +310,7 @@ def run(
             deployment = serve.deployment(name=deployment_name)(import_path)
             deployments = [deployment]
 
-            ray.init(address=cluster_address, namespace="serve")
             serve.start(detached=True)
-            ServeSubmissionClient(dashboard_address)._upload_working_dir_if_needed(
-                runtime_env_updates
-            )
-
-            configure_runtime_env(deployment, runtime_env_updates)
             deployment.options(
                 init_args=args,
                 init_kwargs=kwargs,
@@ -486,11 +340,8 @@ def run(
 
 
 @cli.command(
-    short_help="[Experimental] Get info about your Serve application's config.",
-    help=(
-        "Prints the configurations of all running deployments in the Serve "
-        "application."
-    ),
+    help="[Experimental] Get info about your Serve application's config.",
+    hidden=True,
 )
 @click.option(
     "--address",
@@ -500,32 +351,18 @@ def run(
     type=str,
     help='Address of the Ray dashboard to query. For example, "http://localhost:8265".',
 )
-@click.option(
-    "--json_format",
-    "-j",
-    is_flag=True,
-    help="Print info as json. If omitted, info is printed as YAML.",
-)
-def info(address: str, json_format=bool):
-
-    app_info = ServeSubmissionClient(address).get_info()
-    if app_info is not None:
-        if json_format:
-            print(json.dumps(app_info, indent=4))
-        else:
-            print(yaml.dump(app_info))
+def info(address: str):
+    full_address_path = f"{address}/api/serve/deployments/"
+    response = requests.get(full_address_path)
+    if response.status_code == 200:
+        print(json.dumps(response.json(), indent=4))
+    else:
+        log_failed_request(response, address)
 
 
 @cli.command(
-    short_help="[Experimental] Get your Serve application's status.",
-    help=(
-        "Prints status information about all deployments in the Serve application.\n\n"
-        "Deployments may be:\n\n"
-        "- HEALTHY: all replicas are acting normally and passing their health checks.\n"
-        "- UNHEALTHY: at least one replica is not acting normally and may not be "
-        "passing its health check.\n"
-        "- UPDATING: the deployment is updating."
-    ),
+    help="[Experimental] Get your Serve application's status.",
+    hidden=True,
 )
 @click.option(
     "--address",
@@ -536,17 +373,17 @@ def info(address: str, json_format=bool):
     help='Address of the Ray dashboard to query. For example, "http://localhost:8265".',
 )
 def status(address: str):
-
-    app_status = ServeSubmissionClient(address).get_status()
-    if app_status is not None:
-        print(json.dumps(app_status, indent=4))
+    full_address_path = f"{address}/api/serve/deployments/status"
+    response = requests.get(full_address_path)
+    if response.status_code == 200:
+        print(json.dumps(response.json(), indent=4))
+    else:
+        log_failed_request(response, address)
 
 
 @cli.command(
-    short_help=(
-        "[EXPERIMENTAL] Deletes all running deployments in the Serve application."
-    ),
-    help="Deletes all running deployments in the Serve application.",
+    help="[Experimental] Get info about your Serve application's config.",
+    hidden=True,
 )
 @click.option(
     "--address",
@@ -558,7 +395,6 @@ def status(address: str):
 )
 @click.option("--yes", "-y", is_flag=True, help="Bypass confirmation prompt.")
 def delete(address: str, yes: bool):
-
     if not yes:
         click.confirm(
             f"\nThis will shutdown the Serve application at address "
@@ -567,8 +403,11 @@ def delete(address: str, yes: bool):
             abort=True,
         )
 
-    ServeSubmissionClient(address).delete_application()
-
-    cli_logger.newline()
-    cli_logger.success("\nSent delete request successfully!\n")
-    cli_logger.newline()
+    full_address_path = f"{address}/api/serve/deployments/"
+    response = requests.delete(full_address_path)
+    if response.status_code == 200:
+        cli_logger.newline()
+        cli_logger.success("\nSent delete request successfully!\n")
+        cli_logger.newline()
+    else:
+        log_failed_request(response, address)
