--- python/ray/serve/tests/test_config_files/fake_runtime_env.yaml
@@ -1,21 +1,0 @@
-deployments:
-
-  - name: one
-    init_args: null
-    init_kwargs: null
-    import_path: "test_module.test.one"
-    num_replicas: 2
-    route_prefix: "/one"
-    max_concurrent_queries: null
-    user_config: null
-    autoscaling_config: null
-    graceful_shutdown_wait_loop_s: null
-    graceful_shutdown_timeout_s: null
-    health_check_period_s: null
-    health_check_timeout_s: null
-    ray_actor_options: 
-      runtime_env:
-        py_modules:
-          - "https://fakemodule1.com"
-          - "https://fakemodule2.com"
-        working_dir: "https://fakewd1.com"
