--- python/ray/serve/tests/test_config_files/scarlet.yaml
@@ -1,18 +1,0 @@
-deployments:
-
-  - name: Macaw
-    init_args:
-      - "red"
-    init_kwargs:
-      name: "Scarlet"
-    import_path: "test_cli.Macaw"
-    num_replicas: 1
-    route_prefix: "/Scarlet"
-    max_concurrent_queries: null
-    user_config: null
-    autoscaling_config: null
-    graceful_shutdown_wait_loop_s: null
-    graceful_shutdown_timeout_s: null
-    health_check_period_s: null
-    health_check_timeout_s: null
-    ray_actor_options: null
